Start to add a code editor
not sure if this is good or bad. The idea is to turn
all the inline examples into codepen like things so you can
see the source without having to right-click (like why is that
hard I have no idea).

In any case this is the start of a proof of concept. As usually
my CSS sucks and I don't understand why I can't get a simple
2 panes each using 50% of the space. Will I ever understand CSS?

Otherwise the goal (maybe a bad goal) is that the examples
should run standalone as they are now. In other words no
code should change to run them outside the editor.

Also, no server needed.

So, the current solution is there's an editor at /webgl/resources/editor.html.
The example embedder template `build/templates/example.template` inserts links
like this

    <iframe src="/webgl/resources/editor.html?url=/webgl/name-of-example.html"></iframe>

`editor.js` looks a that URL and uses ajax to load the file. It then creates a blob from
the string and points an iframe to the blob.

This works. In other words it's possible to make a editor that doesn't need a server to
serve the iframe - except relative paths to external resources don't work because they end
up being relative to the blob.

That's mostly ok I guess? I can fix various paths except paths to images embeded in code.
I can hack those as well I suppose. Is that the correct solution? I don't want to hard
code paths since thent they won't run in dev.

Other issues, should I show the entire file in the editor or break it into HTML/JS/CSS
like many other editors? Also the CSS is boilerplate so maybe that should not be shown?
Do I need to show the HTML? Do I show all HTML or try to simplify?

Do I want to show the script tags in HTML for the external libraries?